- Removed resolve bar mechanic - Replaced with new resolve mechanic - Added an End Turn button

+++ Demo/Assets/Character/EnemyCharactersLibrary/EnemyCharacter.cs
@@ -75,13 +75,16 @@ public abstract class EnemyCharacter : Character
 	
 	public void GetStunned()
 	{
-		EnemyMove EM = Moves.Peek();
-		//Pop the move onto the stack without removing
-		if (!EM.IsSpecial())
+		if (Moves.Count > 0)
 		{
-			EM = Moves.Pop();
-			EM.DeleteMove();
-		}			
+			EnemyMove EM = Moves.Peek();
+			//Pop the move onto the stack without removing
+			if (!EM.IsSpecial())
+			{
+				EM = Moves.Pop();
+				EM.DeleteMove();
+			}	
+		}		
 		Moves.Push(new EnemyStunnedMove());
 		this.IsStunned = true;
 		StunnedBuff B = new StunnedBuff(this, this, null, 1);
